Check the result of the cwd in SshShell (#48)
* Check the result of the cwd in SshShell

Using the cwd argument used to run something like that:
$ cd '/some/path'; command -v 'ls' > /dev/null 2>&1 ||
  which 'ls' > /dev/null 2>&1; echo $?; exec 'ls'

Even if the directory doesn't exist, the command is run.
I first tried to change it to:
$ cd '/some/path' && { command -v 'ls' > /dev/null 2>&1 ||
  which 'ls' > /dev/null 2>&1; echo $?; exec 'ls'; }

But we can't check the which/command result if the cd fails. To minimize
the impact, I changed it to:
$ cd '/some/path' 2>/dev/null; command -v 'ls' > /dev/null 2>&1 ||
  which 'ls' > /dev/null 2>&1; echo $?; cd '/some/path' && exec 'ls'

We have to run cd before to check if the command exists because we might
run a script in the cwd directory.
So we don't catch if this cd fails and run it again just before the exec
where we catch it. The only drawback is that if a command is not found
because the cwd is incorrect, we'll raise NoSuchCommandError instead of
RunProcessError.

Fixes #46

* Raise NoSuchDirectoryError if cwd fails

Add new exception "NoSuchDirectoryError".
The exception is raised if trying to change to a non-existent directory.
The behaviour is consistent between local and ssh shell.