Do not leave python 2 blocked while SSH processes run (fixes #61)
If we wait for an SSH process in Python 2, there is no way to cancel the
script unless we send SIGKILL to it (e.g., SIGINT won't work).

The following example will never raise KeyboardInterrupt in Python 2 until
the process finishes:

  import spur
  s = spur.SshShell(hostname="localhost",
                    missing_host_key=spur.ssh.MissingHostKey.accept)
  s.run(["bash", "-c", "sleep 100000"])

We instead add an active wait loop (only in Python 2) so that it can catch
up with queued signals.